// pharos-ai-nextjs/src/types/domain.ts
export type Significance = 'BREAKING' | 'HIGH' | 'STANDARD';
export type AccountType = 'military' | 'government' | 'journalist' | 'analyst' | 'official';

export interface XPost {
  id: string;
  handle: string;
  displayName: string;
  avatar: string;
  avatarColor: string;
  verified: boolean;
  accountType: AccountType;
  significance: Significance;
  timestamp: string;
  content: string;
  images?: string[];
  videoThumb?: string;
  likes: number;
  retweets: number;
  replies: number;
  views: number;
  eventId?: string;
  actorId?: string;
  pharosNote?: string;
}